composition: move strings inside subgraphs
It makes the crate organization cleaner, clearer.
