+++ crates/composition/src/compose.rs
@@ -7,7 +7,7 @@ mod object;
 pub(crate) use self::context::Context as ComposeContext;
 
 use self::{context::Context, input_object::*};
-use crate::subgraphs::{DefinitionKind, DefinitionWalker, FieldWalker};
+use crate::subgraphs::{DefinitionKind, DefinitionWalker, FieldWalker, StringId};
 use itertools::Itertools;
 
 pub(crate) fn compose_subgraphs(ctx: &mut Context<'_>) {
