+++ crates/composition/src/compose/input_object.rs
@@ -1,5 +1,3 @@
-use crate::strings::StringId;
-
 use super::*;
 use std::collections::{HashMap, HashSet};
