+++ crates/composition/src/compose/object.rs
@@ -1,5 +1,5 @@
 use super::*;
-use crate::{strings::StringId, subgraphs::FieldTypeId};
+use crate::subgraphs::{FieldTypeId, StringId};
 use std::collections::HashSet;
 
 /// The arguments of a federated graph's fields are the interseciton of the subgraph's arguments for
